// client/typechain-types/NFT.ts
export type NftMintedEvent = TypedEvent<
  [string, BigNumber, string, string],
  {
    nftContractAddress: string;
    tokenId: BigNumber;
    mintedTo: string;
    tokenUri: string;
  }
>;